Make ModulesFileModelGenerator and associated classes generic.
This is inline with the previous conversion of ModelGenerator to generic form.
As there, existing use of ModulesFileModelGenerator should be largely
unaffected, since it is set up to work using doubles by default.
In particular, the existing constructors continue to work. However,
for compatibility going forward, it is preferable to now use the static
ModulesFileModelGenerator.create(...) method instead.

The "associated classes" are Updater, TransitionList, Choice
and ChoiceListFlexi. The two classes ChoiceList and ChoiceSingleton
have been removed since they are currently unused and out of date.