Replace ModulesFileModelGeneratorSymbolic with generic version.
ModulesFileModelGeneratorSymbolic is used for building models for
the parametric model checking engine. Its now replaced with instances
of ModulesFileModelGenerator<Function>, using the new generic classes.

Various classes in the param package, which heavily overlapped with
the corresponding classes in the simulator package are now removed:
ChoiceListFlexi, TransitionList, SymbolicEngine, ModelGeneratorSymbolic.

In some related re-arranging, conversion of PRISM expressions to
(parametric) Function objects is moved into FunctionFactory from
ModelBuilder (which will eventually be replaced by ConstructModel).
Lookups in ConstantLists, previously done there, are now done in
ModulesFileModelGenerator, where the rest of constant replacement is.

Code to compute choice probability sums and normalise is also
removed from ModelBuilder since this is now handled in the
ModulesFileModelGeneratorSymbolic class, in the same fashion as
for non-parametric models.