Make RewardGenerator generic: rewards can be any type.
Classes that implement this interface using the normal double type
should still work with minimal changes, namely changing some methods
return types from double to Double, e.g. getStateReward.
Ideally they should also specify "implements RewardGenerator<Double>"
instead of "implements ModelGenerator", but this is not essential.

In a related change, default method implementations in RewardGenerator
now throw an exception if a reward should be defined but is not,
e.g., if the number of reward structures has been specified to be >0.
Previously, they returned a default of 0.0, which is no longer suitable
for any reward type (and probably not sensible anyway).