Convert explicit engine model and reward classes to generic types.
All classes and interfaces for models and rewards in the explicit engine
are now generic, with the type variable (Value) denoting the type used
to represent probabilities/rewards/etc. All model classes have a
corresponding Evaluator attached, which can be used to perform
operations on that type. This is available via the method
getEvaluator(). For example, getEvaluator().isZero(x) can be used to
check if Value x is zero,

A default implementation of getEvaluator() in the top-level interface
Model creates a fresh Evaluator for doubles to simplify usage in this
common case. The base class ModelExplicit provides storage for a custom
Evaluator which can be set with setEvaluator(). This is also initialised
to a fresh Evaluator for doubles by default. The same approach is taken
in the Rewards interface, and a new RewardsExplicit superclass is added
to allow an Evaluator to be stored. The Distribution class, used by some
models, is also now generic and also stores a reference to an Evaluator.

Most model classes and most of their methods are now converted to use
the generic type variable instead of doubles. One exception is methods
such as mvMult (multiplication with a vector of doubles), used e.g. in
value iteration and other numerical methods. These continue to assume
the use of doubles. The DTMCSparse and MDPSparse classes also only work
with doubles. MDPGeneric is now redundant and has been removed.

Various bits of functionality that manipulate models have also been
converted to use the generic type variable instead of doubles. Most
notably, this includes explicit state model construction
(ConstructModel). Passing a ModelGenerator<Value> yields a Model<Value>,
where, as usual, the type of Model depends on the type specified in the
ModelGenerator. Construction of rewards (ConstructRewards) is also
supported. Currently, neither of these are actually connected to code
that uses anything apart from doubles.

Various other pieces of code have been converted, such as: export of
state rewards to a file (in ProbModelChecker), LTL product construction
(in LTLModelChecker) and bisimulation minimisation (in Bisimulation).
Some code still assumes the use of doubles, most notably the model
checker classes DTMCModelChecker, MDPModelChecker, etc.. These have been
rewritten (Model -> Model<Double>) to emphasise this, even though this
is mostly only required to fix compiler warnings not errors. Other code
that does not need to know the Value for a model has also mostly been
tidied up, e.g., Model -> Model<?> and new Model() -> new Model<>(), to
remove compiler warnings.